+++ src/SMAPI/Framework/Serialization/Vector2Converter.cs
@@ -6,12 +6,12 @@ using StardewModdingAPI.Toolkit.Serialization.Converters;
 
 namespace StardewModdingAPI.Framework.Serialization
 {
-    /// <summary>Handles deserialization of <see cref="Point"/> for crossplatform compatibility.</summary>
+    /// <summary>Handles deserialization of <see cref="Vector2"/> for crossplatform compatibility.</summary>
     /// <remarks>
     /// - Linux/Mac format: { "X": 1, "Y": 2 }
     /// - Windows format:   "1, 2"
     /// </remarks>
-    internal class PointConverter : SimpleReadOnlyConverter<Point>
+    internal class Vector2Converter : SimpleReadOnlyConverter<Vector2>
     {
         /*********
         ** Protected methods
@@ -19,25 +19,25 @@ namespace StardewModdingAPI.Framework.Serialization
         /// <summary>Read a JSON object.</summary>
         /// <param name="obj">The JSON object to read.</param>
         /// <param name="path">The path to the current JSON node.</param>
-        protected override Point ReadObject(JObject obj, string path)
+        protected override Vector2 ReadObject(JObject obj, string path)
         {
-            int x = obj.ValueIgnoreCase<int>(nameof(Point.X));
-            int y = obj.ValueIgnoreCase<int>(nameof(Point.Y));
-            return new Point(x, y);
+            float x = obj.ValueIgnoreCase<float>(nameof(Vector2.X));
+            float y = obj.ValueIgnoreCase<float>(nameof(Vector2.Y));
+            return new Vector2(x, y);
         }
 
         /// <summary>Read a JSON string.</summary>
         /// <param name="str">The JSON string value.</param>
         /// <param name="path">The path to the current JSON node.</param>
-        protected override Point ReadString(string str, string path)
+        protected override Vector2 ReadString(string str, string path)
         {
             string[] parts = str.Split(',');
             if (parts.Length != 2)
-                throw new SParseException($"Can't parse {typeof(Point).Name} from invalid value '{str}' (path: {path}).");
+                throw new SParseException($"Can't parse {typeof(Vector2).Name} from invalid value '{str}' (path: {path}).");
 
-            int x = Convert.ToInt32(parts[0]);
-            int y = Convert.ToInt32(parts[1]);
-            return new Point(x, y);
+            float x = Convert.ToSingle(parts[0]);
+            float y = Convert.ToSingle(parts[1]);
+            return new Vector2(x, y);
         }
     }
 }